Mapping and Restructuring of code
