+++ catalogue_graph/tests/adapters/steps/axiell_folio_sync/test_deletion_fact_contract.py
@@ -8,11 +8,12 @@ source_id derivation.
 
 from datetime import UTC, datetime
 
-from adapters.steps.axiell_folio_sync.mapper import extract, parse_xml
 from adapters.steps.axiell_folio_sync.mapping import (
     _holdings_hrid,
     _instance_hrid,
     _item_hrid,
+    extract,
+    parse_xml,
 )
 from adapters.transformers.builders.axiell_work_builder import AxiellWorkBuilder
 from adapters.utils.axiell_changeset_reader import SupersededGuid
